Add end to end testing and cleanup

+++ Actions/AL-Go-Helper.ps1
@@ -1172,6 +1172,9 @@ function CheckAndCreateProjectFolder {
             $settingsJson = Get-Content $ALGoSettingsFile -Encoding UTF8 | ConvertFrom-Json
             if ($settingsJson.appFolders.Count -eq 0 -and $settingsJson.testFolders.Count -eq 0) {
                 OutputWarning "Converting the repository to a multi-project repository as no other apps have been added previously."
+                if (!(Test-Path "AL-Go-Settings.json")) {
+                    Copy-Item $ALGoSettingsFile -Destination ".github\AL-Go-Settings.json"
+                }
                 New-Item $project -ItemType Directory | Out-Null
                 Move-Item -path $ALGoFolder -Destination $project
                 Set-Location $project
@@ -1182,15 +1185,20 @@ function CheckAndCreateProjectFolder {
         }
         else {
             if (!(Test-Path $project)) {
-                OutputWarning "Project folder doesn't exist, creating a new project folder and a default settings file with type PTE and country us. Please modify if needed."
                 New-Item -Path (Join-Path $project $ALGoFolder) -ItemType Directory | Out-Null
                 Set-Location $project
-                [ordered]@{
-                    "type" = "PTE"
-                    "country" = "us"
-                    "appFolders" = @()
-                    "testFolders" = @()
-                } | ConvertTo-Json | Set-Content $ALGoSettingsFile -Encoding UTF8
+                if (Test-Path "..\.github\AL-Go-Settings.json") {
+                    Copy-Item -Path "..\.github\AL-Go-Settings.json" -Destination $ALGoSettingsFile
+                }
+                else {
+                    OutputWarning "Project folder doesn't exist, creating a new project folder and a default settings file with type PTE and country us. Please modify if needed."
+                    [ordered]@{
+                        "type" = "PTE"
+                        "country" = "us"
+                        "appFolders" = @()
+                        "testFolders" = @()
+                    } | ConvertTo-Json | Set-Content $ALGoSettingsFile -Encoding UTF8
+                }
             }
             else {
                 Set-Location $project
